perf(pap): establish Triton decode milestone
Switch the P17 attention main path from paged FlashAttention to a fixed split-4 Triton decode kernel and reuse one workspace across a decode step. Keep trace compatibility and add exact-shape diagnostics plus focused runtime coverage.

The validated C4 steady TPOT improves from 49.75 ms to 42.47 ms with cache, routing, async-token, MPS, drain, and output correctness audits passing.

+++ vllm/v1/attention/ops/triton_decode_attention.py
@@ -127,13 +127,14 @@ def _fwd_kernel_stage1(
         vs = tl.load(v_scale)
         for start_n in range(split_kv_start, split_kv_end, BLOCK_N):
             offs_n = start_n + tl.arange(0, BLOCK_N)
+            # Cross-layer KV strides can make page offsets exceed int32.
             kv_page_number = tl.load(
                 Req_to_tokens
                 + stride_req_to_tokens_b * cur_batch_req_idx
                 + offs_n // PAGE_SIZE,
                 mask=offs_n < split_kv_end,
                 other=0,
-            )
+            ).to(tl.int64)
             kv_in_page = offs_n % PAGE_SIZE
             offs_buf_k = (
                 (kv_page_number * stride_buf_kpbs + kv_in_page * stride_buf_kbs)[
@@ -368,6 +369,7 @@ def _fwd_grouped_kernel_stage1(
         vs = tl.load(v_scale)
         for start_n in tl.range(split_kv_start, split_kv_end, BLOCK_N):
             offs_n = start_n + tl.arange(0, BLOCK_N)
+            # Cross-layer KV strides can make page offsets exceed int32.
             kv_page_number = tl.load(
                 Req_to_tokens
                 + stride_req_to_tokens_b * cur_batch_req_idx
@@ -375,7 +377,7 @@ def _fwd_grouped_kernel_stage1(
                 mask=offs_n < split_kv_end,
                 other=0,
                 cache_modifier=".ca",
-            )
+            ).to(tl.int64)
             kv_off_k = (
                 kv_page_number * stride_buf_kpbs + (offs_n % PAGE_SIZE) * stride_buf_kbs
             )
